Collect the T5 graph regression test
Move the deterministic standard-T5 graph contract assertion into the pytest-discovered co-located T5 test module so CI exercises the 176-node contract, component counts, decoder input set, and historical cross-cache shape.
